+++ pyppeteer/launcher.py
@@ -24,6 +24,7 @@ from pyppeteer import __pyppeteer_home__
 from pyppeteer.browser import Browser
 from pyppeteer.connection import Connection
 from pyppeteer.errors import BrowserError
+from pyppeteer.helper import debugError
 from pyppeteer.util import check_chromium, chromium_excutable
 from pyppeteer.util import download_chromium, merge_dict, get_free_port
 
@@ -210,14 +211,14 @@ class Launcher(object):
 
     async def killChrome(self) -> None:
         """Terminate chromium process."""
-        logger.debug('terminate chrome process...')
+        logger.info('terminate chrome process...')
         if self.connection and self.connection._connected:
             try:
                 await self.connection.send('Browser.close')
                 await self.connection.dispose()
-            except Exception:
+            except Exception as e:
                 # ignore errors on browser termination process
-                pass
+                debugError(logger, e)
         if self._tmp_user_data_dir and os.path.exists(self._tmp_user_data_dir):
             # Force kill chrome only when using temporary userDataDir
             self.waitForChromeToClose()